# api-docs.yaml
openapi: 3.0.1
info:
  title: Book Checkout Store
  description: Assignment
  termsOfService:

  version: "1.0"
servers:
  - url: "http://localhost:8080"
    description: Generated server url
paths:
  /api/book:
    get:
      tags:
        - book-controller
      operationId: getAllBook
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDTO"
    post:
      tags:
        - book-controller
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookDTO"
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: object
  /api/book/{id}:
    get:
      tags:
        - book-controller
      operationId: getBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookDTO"
    put:
      tags:
        - book-controller
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookDTO"
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
        - book-controller
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: default response
  /api/checkout:
    post:
      tags:
        - checkout-controller
      operationId: checkout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequestDTO"
      responses:
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CheckoutResponseDTO"
components:
  schemas:
    BookDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        author:
          type: string
        type:
          type: string
        price:
          type: number
          format: double
        isbn:
          type: string
    CheckoutRequestDTO:
      type: object
      properties:
        checkoutList:
          type: array
          checkoutList:
            bookId: integer
            quantity: integer
            format: int64
        promoCode:
          type: string
    CheckoutResponseDTO:
      type: object
      properties:
        total:
          type: number
          format: double